ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
shap==0.43.0
pillow==10.1.0
tensorflow==2.15.0
python-multipart==0.0.6
joblib==1.3.2

